Moved data.py contents to new thematic submodules
