+++ PyUVS/__init__.py
@@ -1,7 +1,10 @@
-from . import data
+from . import search
+from . import download
 from . import geometry
 from . import graphics
 from . import miscellaneous
 from . import spice
 from . import statistics
 from . import time
+from . import instrument
+from . import constants
