+++ PyUVS/constants.py
@@ -1,0 +1,3 @@
+"""Physical constants used in multiple locations."""
+
+R_Mars_km = 3.3895e3  # [km]
